+++ arviz/preview.py
@@ -1,0 +1,17 @@
+# pylint: disable=unused-import,unused-wildcard-import,wildcard-import
+"""Expose features from arviz-xyz refactored packages inside ``arviz.preview`` namespace."""
+
+try:
+    from arviz_base import *
+except ModuleNotFoundError:
+    pass
+
+try:
+    import arviz_stats
+except ModuleNotFoundError:
+    pass
+
+try:
+    from arviz_plots import *
+except ModuleNotFoundError:
+    pass